fixed missing top level case for editing choice requirements
